Refactor inserter button as entry point
