+++ languages/gutenberg.pot
@@ -38,11 +38,11 @@ msgstr ""
 msgid "Justify"
 msgstr ""
 
-#: editor/components/inserter/button.js:31
+#: editor/components/inserter/index.js:31
 msgid "Add a block"
 msgstr ""
 
-#: editor/components/inserter/index.js:40
+#: editor/components/inserter/menu.js:40
 msgid "Search…"
 msgstr ""
